# attribute: faster NewSet, NewSetWithFiltered