@dmathieu 2025-03-07
Could you use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to make the benchmark more comparable ?

You also need to sign the CLA