@akehlenbeck 2025-03-07
> Could you use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to make the benchmark more comparable ?

Yes, the output I posted did come from benchstat, yes. Is there some other formatting mode you typically use with it?

> You also need to sign the CLA

Yup, I'm waiting on my company's approvers. Will the CLA status here update automatically when they approve me?