@akehlenbeck 2025-12-15
I have been unable to find someone at my company who can approve the CLA for me... I'm going to close this change just so that it stops cluttering up my list of open PRs, but you are more than welcome to take the code and re-author it if you would like...